requirements.txt
fastapi
uvicorn
pillow
python-multipart
torch
diffusers
transformers
ultralytics
fpdf
smtplib
email-validator
authlib
python-dotenv
